refactoring to general project
some stuff is a work in progress
