+++ config_templates/api-webapp-data-sources.xml
@@ -20,31 +20,41 @@ xsi:schemaLocation="
     <property name="password" value="{pg_password}" />
 </bean>
 
+<!-- Geographic center of system -->
 <bean id="externalGeocoderImpl" class="org.onebusaway.geocoder.impl.FixedGeocoderImpl">
-    <property name="lat" value="42.32451" />
-    <property name="lon" value="-122.8717556" />
-    <property name="city" value="Medford" />
-    <property name="state" value="OR" />
-    <property name="postalCode" value="97501" />
+    <property name="lat" value="{geocode_center_lat}" />
+    <property name="lon" value="{geocode_center_lon}" />
+    <property name="city" value="{geocode_center_city}" />
+    <property name="state" value="{geocode_center_state}" />
+    <property name="postalCode" value="{geocode_center_zip}" />
 </bean>
 
 <!-- Service that is used to verify API keys -->
 <bean id="apiKeyValidationService" class="org.onebusaway.users.impl.validation.KeyValidationServiceImpl" />
 
+<!-- Official OBA Android App API Key -->
 <bean class="org.onebusaway.users.impl.CreateApiKeyAction">
  	<property name="key" value="v1_BktoDJ2gJlu6nLM6LsT9H8IUbWc=cGF1bGN3YXR0c0BnbWFpbC5jb20="/>
 </bean>
 
+<!-- Official OBA iPhone App API Key -->
 <bean class="org.onebusaway.users.impl.CreateApiKeyAction">
  	<property name="key" value="org.onebusaway.iphone"/>
 </bean>
 
+<!-- Official OBA Windows Phone/Fire Phone App API Key -->
 <bean class="org.onebusaway.users.impl.CreateApiKeyAction">
  	<property name="key" value="v1_C5+aiesgg8DxpmG1yS2F/pj2zHk=c3BoZW5yeUBnbWFpbC5jb20=="/>
 </bean>
 
+<!-- Official OBA Windows 8 App API Key -->
 <bean class="org.onebusaway.users.impl.CreateApiKeyAction">
  	<property name="key" value="693c0a55-9ef0-4302-8bc3-f9b2db93e124"/>
+</bean>
+
+<!-- Official OBA Alexa (for Amazon Echo, etc.) App API Key -->
+<bean class="org.onebusaway.users.impl.CreateApiKeyAction">
+ 	<property name="key" value="e5c0e97a-729d-4fdb-a3ca-2fccb20ac3ab"/>
 </bean>
 
 {api_testing}
