+++ config_templates/aws.ini
@@ -1,5 +1,5 @@
 [DEFAULT]
-ami_id = ami-e7527ed7
+ami_id = {ami_id}
 aws_access_key_id = {aws_access_key}
 aws_secret_access_key = {aws_secret_key}
 client_name = {client_name}
