+++ config_templates/oba.ini
@@ -1,8 +1,17 @@
 [DEFAULT]
 allow_api_test_key = {allow_api_test_key}
-oba_base_folder = onebusaway-application-modules-rvtd
-oba_git_branch = rvtd-1.1.13.install
-oba_git_repo = https://github.com/trilliumtransit/onebusaway-application-modules-rvtd.git
+oba_base_folder = {oba_base_folder}
+oba_git_branch = {oba_git_branch}
+oba_git_repo = {oba_git_repo}
 pg_username = {pg_username}
 pg_password = {pg_password}
 pg_role = {pg_role}
+geocode_center_lat = {geocode_center_lat}
+geocode_center_lon = {geocode_center_lon}
+geocode_center_city = {geocode_center_city}
+geocode_center_state = {geocode_center_state}
+geocode_center_zip = {geocode_center_zip}
+service_area_bounds_min_lat = {service_area_bounds_min_lat}
+service_area_bounds_min_lon = {service_area_bounds_min_lon}
+service_area_bounds_max_lat = {service_area_bounds_max_lat}
+service_area_bounds_max_lon = {service_area_bounds_max_lon}
