+++ config_templates/watchdog.ini
@@ -7,7 +7,7 @@ end_of_day = 74700
 from_address = {from_address}
 imap_host = imap.gmail.com
 password = {password}
-realtime_agencies = 1739
+realtime_agencies = {realtime_agencies}
 report_hours = 5,15
 report_list = {report_list}
 report_status_file = {report_status_file}
